backend_requirements.txt
# FastAPI and API server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database dependencies
psycopg2-binary>=2.9.8
sqlalchemy>=2.0.25

# Redis dependencies
redis[asyncio]>=5.0.1

# Async and HTTP dependencies
httpx>=0.25.0
aiofiles>=23.2.0

# Configuration and validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Additional utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Production server dependencies
gunicorn>=21.2.0

# Browser automation dependencies
playwright>=1.40.0


